during expiry, handle the situation where the newest tokens is in the future by resetting the newest token stamp to current time, and changing any token atimes > newest atime to newest atime.
git-svn-id: https://svn.apache.org/repos/asf/incubator/spamassassin/trunk@10185 13f79535-47bb-0310-9956-ffa450edef68

+++ lib/Mail/SpamAssassin/BayesStore/SQL.pm
@@ -250,6 +250,15 @@ sub token_expiration {
   # Figure out how old is too old...
   my $too_old = $vars[10] - $newdelta; # tooold = newest - delta
 
+  # if token atime > newest, reset to newest ...
+  my $sql = "UPDATE bayes_token SET atime=? WHERE username = ? and atime > ?";
+  my $rows = $self->{_dbh}->do($sql, undef, $vars[10], $self->{_username}, $vars[10]);
+  unless (defined($rows)) {
+    dbg("bayes: reset tokens in future: SQL Error: ".$self->{_dbh}->errstr());
+    return 0;
+  }
+
+  # Do the expire
   my $sql = "DELETE from bayes_token WHERE username = ? and atime < ?";
 
   my $rows = $self->{_dbh}->do($sql, undef, $self->{_username}, $too_old);
